feat(JobCard): Differentiate brand color of the branded ad stripe (#392)
Currently JobCard always renders the branded ad with DB color (any brand, including Jobstreet). Wrapping it here in the styleguide context that allows us to differentiate db/jobstreet for now and render it with the different color.

This is for sure a temporary solution till we properly implement the things via braid and shared ui

To test this you'll need to switch the default brand in the styleguide provider in the app to jobstreet or jobsdb instead of seekasia.